Specify error statuses on wp_die messages.
The default response is 500 server error when using wp_die, but some of these error messages are being used for "Download does not exist" (should be 404 as it's a missing file) and "You do not have permission" (should be 200 as it's a user error & has proper messaging for informing the user of their requirements).

I noticed Google Webmaster Tools was reporting a few member only files as 500 server errors, and it spurred me to giving this a look. I should mention that I left the "No file paths defined" error as the default 500 server error response.