Add template_dir attribute to shortcodes
This enables users to specify a custom directory containing templates
in shortcodes. Some themes wipe-out their entire theme directory on
update, so this allows for the possibility of a more-permanent
location that will not be cleared when either this plugin or a theme
is updated.

The template_dir attribute should be a directory relative to
WordPress' home URL, e.g. wp-content/dlm-templates