Apply double-counting log prevention to version_id
The wp_dlm_downloading cookie currently does nothing. There is no
check in place to actually reference the value in the cookie and
prevent logging. Fix this in the log() function.

Distinct versions of the same download should still both log within
the 60sec timeout. Change the contents of the wp_dlm_downloading
cookie to version_id instead of the overall download id.