Handle logging/counting functionality in log()
The current implementation of the wp_dlm_downloading cookie cooldown
only applies to download counts. This can lead to inconsistencies
between the number of log entries created and the number of downloads
counted. Expand the functionality of log() (which can be renamed later
if desired) to:

* Perform 60sec wp_dlm_downloading cookie cooldown test and set
* Write log and increment download count based on unique IP option (if
  logging is enabled)

This brings back the original intention of the cookie cooldown to
reduce redundant entries in the log for ill-behaved download managers
and anxious clickers.