Logs: Fix display for unknown user in exported log
When generating rows to be exported to a csv log, the $user variable
is retained between loop iterations. Then, if a subsequent log entry
does not have user information, the user from the previous iteration
will be applied, so it appears as if a user has downloaded more than
that for which they are actually responsible. Clear the $user variable
after usage.