Make the permalinks for downloads use friendly URLs
With `rewrite` set to `false` on both dlm_download and dlm_download_version post types, the resulting download URL that's surfaced to the site (and site search) is something like `?dlm_download=_download-slug_`. Meanwhile, setting `rewrite` to `true` makes it use the endpoint that's set in Download Monitor's settings/options.